+++ collect-webapp/frontend/src/datamanagement/components/recordeditor/Table/Table.js
@@ -13,89 +13,12 @@ import DeleteIconButton from 'common/components/DeleteIconButton'
 
 import EntityCollectionComponent from '../EntityCollectionComponent'
 import FormItemFieldComponent from '../FormItemFieldComponent'
-import * as FieldsSizes from '../fields/FieldsSizes'
 import HeadingRow from './HeadingRow'
+
+import * as TableColumns from './tableColumns'
 
 const ROW_NUMBER_COLUMN_WIDTH = 60
 const DELETE_COLUMN_WIDTH = 60
-
-const calculateWidth = (headingColumn) => FieldsSizes.getWidth({ fieldDef: headingColumn, inTable: true })
-
-const determineColumnInfo = ({ headingColumn, entities, col }) => {
-  const { attributeDefinition, col: colOriginal, colSpan } = headingColumn
-  const { alwaysRelevant, hideWhenNotRelevant, id: attributeDefinitionId, calculated, hidden } = attributeDefinition
-
-  const relevant =
-    alwaysRelevant ||
-    entities.some((entity) => entity.hasSomeDescendantRelevant({ nodeDefinition: attributeDefinition }))
-
-  const isNotEmpty = () =>
-    entities.length > 0 &&
-    entities.some((entity) => entity.hasSomeDescendantNotEmpty({ nodeDefinition: attributeDefinition }))
-
-  const visible = !(calculated && hidden) && (!hideWhenNotRelevant || relevant || isNotEmpty())
-
-  return {
-    colOriginal,
-    col: visible ? col : -1,
-    colSpan,
-    attributeDefinitionId,
-    relevant,
-    visible,
-  }
-}
-
-const determineColumnInfoByAttributeDefinitionId = ({ entities, itemDef }) => {
-  const { headingColumns } = itemDef
-
-  let currentCol = 1
-  return headingColumns.reduce((infoByDefId, headingColumn) => {
-    const { attributeDefinition } = headingColumn
-    const info = determineColumnInfo({ headingColumn, entities, col: currentCol })
-    infoByDefId[attributeDefinition.id] = info
-    if (info.visible) {
-      currentCol = info.col + headingColumn.colSpan
-    }
-    return infoByDefId
-  }, [])
-}
-
-const determineColumnsVisible = ({ itemDef, columnInfoByDefId }) => {
-  const { headingColumns } = itemDef
-
-  return headingColumns.reduce((columnsVisible, headingColumn) => {
-    const { attributeDefinition } = headingColumn
-
-    if (columnInfoByDefId[attributeDefinition.id].visible) {
-      columnsVisible.push(headingColumn)
-    }
-    return columnsVisible
-  }, [])
-}
-
-const calculateCol = (columnsInfo) => (headingColumn) => {
-  const columnsInfoArr = Object.values(columnsInfo).sort((infoA, infoB) => infoA.colOriginal - infoB.colOriginal)
-
-  const firstVisibleColumn =
-    headingColumn instanceof ColumnDefinition
-      ? headingColumn
-      : // if headingColumn is a group, find the first visible descendant (attribute) column
-        headingColumn.descendantColumns.find(
-          (descendantColumn) => columnsInfo[descendantColumn.attributeDefinitionId].visible
-        )
-
-  let col = 1
-  columnsInfoArr.some((info) => {
-    if (info.attributeDefinitionId === firstVisibleColumn.attributeDefinitionId) {
-      return true
-    }
-    if (info.visible) {
-      col += info.colSpan
-    }
-    return false
-  })
-  return col
-}
 
 export default class Table extends EntityCollectionComponent {
   constructor() {
@@ -132,8 +55,8 @@ export default class Table extends EntityCollectionComponent {
     const { showRowNumbers, entityDefinition, headingRows } = itemDef
     const { enumerate } = entityDefinition
 
-    const columnInfoByDefId = determineColumnInfoByAttributeDefinitionId({ entities, itemDef })
-    const columnsVisible = determineColumnsVisible({ itemDef, columnInfoByDefId })
+    const columnInfoByDefId = TableColumns.determineColumnInfoByAttributeDefinitionId({ entities, itemDef })
+    const columnsVisible = TableColumns.determineColumnsVisible({ itemDef, columnInfoByDefId })
     const attributeDefIdsVisible = columnsVisible.map((headingColumn) => headingColumn.attributeDefinition.id)
 
     if (JSON.stringify(columnInfoByDefId) === JSON.stringify(columnInfoByDefIdPrev)) {
@@ -158,7 +81,7 @@ export default class Table extends EntityCollectionComponent {
     )
 
     const headingColumnWidths = columnsVisible.reduce((acc, headingColumn) => {
-      acc.push(calculateWidth(headingColumn))
+      acc.push(TableColumns.calculateWidth(headingColumn))
       return acc
     }, [])
 
@@ -227,7 +150,7 @@ export default class Table extends EntityCollectionComponent {
           <HeadingRow
             key={`heading-row-${index + 1}`}
             columnInfoByDefId={columnInfoByDefId}
-            columnCalculator={calculateCol(columnInfoByDefId)}
+            columnCalculator={TableColumns.calculateColumnPosition(columnInfoByDefId)}
             headingRow={headingRow}
             totalHeadingRows={headingRows.length}
             totalHeadingColumns={headingColumns.length}
@@ -311,7 +234,7 @@ export default class Table extends EntityCollectionComponent {
                 return (
                   <Column
                     key={headingColumn.attributeDefinitionId}
-                    width={calculateWidth(headingColumn)}
+                    width={TableColumns.calculateWidth(headingColumn)}
                     cellRenderer={this.cellRenderer}
                     dataKey={headingColumn}
                     headingColumn={headingColumn}